requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
scikit-learn
python-multipart
requests
pytest
httpx
google-generativeai
